Time: 3 ms (67.08%), Space: 7.1 MB (14.31%) - LeetHub

+++ 8-string-to-integer-atoi/8-string-to-integer-atoi.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int myAtoi(string s) {
-        long long num=0;
+        long int num=0;
         int i=0,sign=0,plus=0,pre=1;
         while(s[i]==' ') i++;
         if(s[i]=='-'){
